Task manager server pat

+++ beeflow/common/tm/clean.sh
@@ -1,6 +1,13 @@
 #! /bin/bash
 
-rm -rf ~/.beeflow/worker/workflow-*
+read -p 'Do you want to remove submitted scripts ~/.beeflow/worker/workflow-* (y/n)?' ans 
+if [ $ans == 'y' ] || [ $ans == 'Y' ] 
+then
+  echo 'Removing scripts!'
+  rm -rf ~/.beeflow/worker/workflow-*
+else
+  echo 'Did not remove scripts.'
+fi
 rm *.log
 rm *.out
 rm *.json
